Big Update. Wrote the entire climbSubsystem and added commands for it as well as started to utilize InstantCommands
Added 2 servos and 2 DoubleSolenoids to climbSubsystem and added in three commands to control these commands (actually going up will be
handled by just driving) the basic sets and toggles I have decided will be used with InstantCommands making life a lot easier. I also
added in an override for the LockPistion. It will attempt to auto lock after the last 1 second of teleop so we dont forget as drivers.